Refactor task sidebar into separate sections
